Split unit-test-backend-8-gpu-amd into matrix partitions (basic and mtp)
