+++ test/srt/run_suite.py
@@ -151,7 +151,8 @@ suites = {
         TestFile("test_disaggregation_dp_attention.py", 155),
         TestFile("test_disaggregation_different_tp.py", 600),
         TestFile("test_disaggregation_pp.py", 140),
-        TestFile("test_full_deepseek_v3.py", 550),
+        TestFile("test_deepseek_v3_basic.py", 275),
+        TestFile("test_deepseek_v3_mtp.py", 275),
     ],
     "per-commit-4-gpu-b200": [
         # TestFile("test_gpt_oss_4gpu.py", 600),
@@ -266,7 +267,8 @@ suite_amd = {
         TestFile("test_pp_single_node.py", 150),
     ],
     "per-commit-8-gpu-amd": [
-        TestFile("test_full_deepseek_v3.py", 250),
+        TestFile("test_deepseek_v3_basic.py", 275),
+        TestFile("test_deepseek_v3_mtp.py", 275),
     ],
     "nightly-amd": [
         TestFile("test_nightly_gsm8k_eval_amd.py"),
@@ -368,7 +370,7 @@ if __name__ == "__main__":
     arg_parser.add_argument(
         "--timeout-per-file",
         type=int,
-        default=1500 if is_hip() else 1200,
+        default=1200,
         help="The time limit for running one file in seconds.",
     )
     arg_parser.add_argument(
